+++ Hammer/Commands/GagCommand.cs
@@ -5,6 +5,7 @@ using DSharpPlus.SlashCommands.Attributes;
 using Hammer.Services;
 using Microsoft.Extensions.Logging;
 using X10D.DSharpPlus;
+using X10D.Time;
 
 namespace Hammer.Commands;
 
@@ -73,7 +74,7 @@ internal sealed class GagCommand : ApplicationCommandModule
     public async Task GagAsync(
         InteractionContext context,
         [Option("user", "The user to gag.")] DiscordUser user,
-        [Option("duration", "The duration of the gag. Defaults to 5 minutes")] TimeSpan? duration = null
+        [Option("duration", "The duration of the gag. Defaults to 5 minutes")] string? duration = null
     )
     {
         var builder = new DiscordEmbedBuilder();
@@ -90,7 +91,14 @@ internal sealed class GagCommand : ApplicationCommandModule
 
         try
         {
-            await _infractionService.GagAsync(user, staffMember, duration: duration).ConfigureAwait(false);
+            if (TimeSpanParser.TryParse(duration, out TimeSpan timeout))
+            {
+                await _infractionService.GagAsync(user, staffMember, duration: timeout).ConfigureAwait(false);
+            }
+            else
+            {
+                await _infractionService.GagAsync(user, staffMember).ConfigureAwait(false);
+            }
 
             builder.WithColor(DiscordColor.Orange);
             builder.WithAuthor(user);