+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug with rule autocompletion (hopefully for the last time).
 - Fixed parsing of /gag duration.
 
+### Changed
+- Temporary ban DM no longer contains ModMail notice.
+
 ### Removed
 - Removed `/migrate` command.