+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.7.0] - 2023-09-06
+
+### Added
+- Added `/rules update` command.
+
 ## [5.6.1] - 2023-07-24
 
 ### Added
@@ -433,6 +438,7 @@ No substantial changes. Commit 3b8259a6cfb82ec0f5f51804c1ac7f1f5880d014 fixed an
 
 - Hammer is released.
 
+[5.7.0]: https://github.com/BrackeysBot/Hammer/releases/tag/v5.7.0
 [5.6.1]: https://github.com/BrackeysBot/Hammer/releases/tag/v5.6.1
 [5.6.0]: https://github.com/BrackeysBot/Hammer/releases/tag/v5.6.0
 [5.5.2]: https://github.com/BrackeysBot/Hammer/releases/tag/v5.5.2